requirements.txt
numpy==1.25.0
pandas==2.1.2    
psycopg2==2.9.7    
boto3==1.28.0   
beautifulsoup4==4.12.2
selenium==4.13.0
apache-airflow-providers-amazon==8.6.0
minio
webdriver-manager
